pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "custom_autocomplete"
version = "2024.02.22"
description = "Simple trie data structure."
authors = [
    "Steven Loscalzo",
]
packages = [{include = "custom_autocomplete", from = "src"}]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "6.0.1"
pandas = "1.5.3"
pyarrow = "15.0.0"
panel = "1.3.8"
